requirements.txt
Flask
Flask-SQLAlchemy
qrcode
python-dotenv
Pillow
gunicorn
psycopg2-binary